backend/pyproject.toml
[project]
name = "cenfotec-gestion-academica-backend"
version = "0.1.0"
description = "Backend API para plataforma de gestión académica"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "celery[redis]>=5.4.0",
    "redis>=5.0.0",
    "minio>=7.2.0",
    "python-multipart>=0.0.9",
    "pgvector>=0.3.0",
    "httpx>=0.27.0",
    "google-genai>=1.0.0",
    "pymupdf>=1.24.0",
    "PyJWT[crypto]>=2.8.0",
    "bcrypt>=4.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0,<1.0.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.6.0",
    "httpx>=0.27.0",
    "testcontainers[postgres]>=4.0.0",
    "psycopg[binary]>=3.2.0",
    "aiosqlite>=0.20.0",
]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
markers = [
    "unit: fast tests without I/O",
    "integration: tests that hit the database",
    "api: end-to-end HTTP tests",
    "requires_postgres: tests that need PostgreSQL (skipped on SQLite)",
]
filterwarnings = [
    "ignore::DeprecationWarning:sqlalchemy.*",
]
